MIDI out: change CC from volume to breath
Previously the midi output would use CC 7 (volume), which doesn't play
nicely with external synthesizers as it changes the channel volume
instead of the note "intensity".

Now, switch to CC 2 (breath) which more closely aligns to the intent
of the command. Also remove code which momentarily dipped the output
value to zero every time the note changed.

NOTE: This change means that synthesizers that do not support breath
control will just receive a `noteon` message with `velocity = 127`
anytime the haxophone reads non-zero breath pressure. Even when not
blowing, the pressure value might oscillate between 0 and 1, causing
frequent notes to sound.

If you are using the fluidsynth CLI as a midi synth, use `setbreathmode`
to enable support for CC2 and fix the above issue.